Commitment has bug
The bug  consist in an incorrect calculating of pi. While executing
algorithm i received  nan  or zero value of pi parameter. But why does
it happen i have no idea yet

+++ kmeanspp.py
@@ -6,7 +6,7 @@ from statistics import *
 
 if __name__ == "__main__":
     kmeans = KMeans(n_clusters=2, init='k-means++')
-    x = np.loadtxt("covvec")
+    x = np.loadtxt("covvec", dtype=int)
     kmeans.fit(x.reshape(-1, 1))    # one-dimensional
 
     # x0, x1 = [], []
